Changing the structural of the Flask project and setting up the project for Vercel deployment
