+++ .gitignore
@@ -1,2 +1,4 @@
 .venv
 node_modules
+__pycache__
+app/__pycache__
