+++ app/templates/pages/about.html
@@ -1,23 +1,12 @@
 {% extends "base.html" %}
 
 {% block content %}
-	<section class="mt-12 mb-8 dark:bg-gray-900">
-		<div class="py-8 px-4 mx-auto max-w-screen-xl lg:py-16">
-			<div class="max-w-4xl mx-auto bg-white dark:bg-gray-800 rounded-lg shadow-md overflow-hidden">
-				<div class="p-6 md:p-10">
-					<h2 class="text-3xl md:text-4xl font-bold text-gray-900 dark:text-white mb-4">About Us</h2>
-					<p class="text-lg text-gray-700 dark:text-gray-400 leading-relaxed mb-6">
-						Welcome to our platform! We provide an innovative tool for predicting unemployment rates and salary potentials based on chosen majors from a rich dataset. With user-friendly features and advanced algorithms, users can input their preferred majors and receive accurate forecasts regarding future job market trends.
-					</p>
-					<p class="text-lg text-gray-700 dark:text-gray-400 leading-relaxed mb-6">
-						Whether you're exploring potential career paths or planning educational pursuits, our platform empowers you with valuable insights to make informed decisions. We leverage data-driven analysis and predictive modeling to provide a valuable resource for individuals navigating the complexities of the modern job market, ultimately aiding in the pursuit of fulfilling and prosperous career opportunities.
-					</p>
-				</div>
-			</div>
+	<section class=" dark:bg-gray-950 mt-16">
+		<div class="py-8 px-4 mx-auto max-w-screen-xl text-center lg:py-16">
+			<h1 class="mb-4 text-4xl font-extrabold tracking-tight leading-none text-gray-900 md:text-5xl lg:text-6xl dark:text-white">Meet our team!</h1>
+			<p class="text-lg font-normal text-gray-500 lg:text-xl sm:px-16 lg:px-48 dark:text-gray-400">Here at Flowbite we focus on markets where technology, innovation, and capital can unlock long-term value and drive economic growth.</p>
 		</div>
 	</section>
-
-	<div class="text-5xl text-center font-bold mb-10">Our Group Project</div>
 
 	<div class="flex justify-center space-x-20">
 		<div class="max-w-sm mb-24 bg-white border border-gray-200 rounded-lg shadow dark:bg-gray-800 dark:border-gray-700">
