+++ app/templates/pages/chart.html
