+++ app/templates/pages/home.html
@@ -1,11 +1,17 @@
 {% extends "base.html" %}
 
 {% block content %}
-    <section class="mt-12 mb-10 dark:bg-gray-900">
+    <section class="mt-12 mb-10 dark:bg-gray-950">
         <div class="py-8 px-4 mx-auto max-w-screen-xl lg:py-16">
             <div class="bg-gray-50 dark:bg-gray-800 border border-gray-200 dark:border-gray-700 rounded-lg p-8 md:p-12 mb-8">
                 <h1 class="text-gray-900 dark:text-white text-3xl md:text-5xl font-extrabold mb-2">Learn more about your future career</h1>
-                <p class="text-lg font-normal text-gray-500 dark:text-gray-400 mb-6">Gain insightful information and valuable projections regarding your prospective career based from unemployment and salary rate.</p>
+                <br>
+                <p class="text-lg text-gray-700 dark:text-gray-400 leading-relaxed mb-6">
+                    Welcome to our platform! We provide an innovative tool for predicting unemployment rates and salary potentials based on chosen majors from a rich dataset. With user-friendly features and advanced algorithms, users can input their preferred majors and receive accurate forecasts regarding future job market trends.
+                </p>
+                <p class="text-lg text-gray-700 dark:text-gray-400 leading-relaxed mb-6">
+                    Whether you're exploring potential career paths or planning educational pursuits, our platform empowers you with valuable insights to make informed decisions. We leverage data-driven analysis and predictive modeling to provide a valuable resource for individuals navigating the complexities of the modern job market, ultimately aiding in the pursuit of fulfilling and prosperous career opportunities.
+                </p>
                 <a href="#" id="scrollButton" class="inline-flex justify-center items-center py-2.5 px-5 text-base font-medium text-center text-white rounded-lg bg-blue-700 hover:bg-blue-800 focus:ring-4 focus:ring-blue-300 dark:focus:ring-blue-900">
                     Get started below 
                     <svg class="w-3.5 h-3.5 ms-2 rtl:rotate-180" aria-hidden="true" xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 14 10">
