+++ requirements.txt
@@ -1,0 +1,26 @@
+blinker==1.8.1
+click==8.1.7
+contourpy==1.2.1
+cycler==0.12.1
+Flask==3.0.3
+fonttools==4.51.0
+itsdangerous==2.2.0
+Jinja2==3.1.3
+joblib==1.4.2
+kiwisolver==1.4.5
+MarkupSafe==2.1.5
+matplotlib==3.8.4
+numpy==1.26.4
+packaging==24.0
+pandas==2.2.2
+pillow==10.3.0
+pyparsing==3.1.2
+python-dateutil==2.9.0.post0
+pytz==2024.1
+scikit-learn==1.4.2
+scipy==1.13.0
+seaborn==0.13.2
+six==1.16.0
+threadpoolctl==3.5.0
+tzdata==2024.1
+Werkzeug==3.0.2
