+++ tailwind.config.js
@@ -5,6 +5,7 @@ module.exports = {
     "./static/src/**/*.js",
     "./node_modules/flowbite/**/*.js"
   ],
+  darkMode: 'class',
   theme: {
     extend: {
       // colors: {
