--- templates/pages/salary.html
@@ -1,52 +1,0 @@
-{% extends "base.html" %}
-
-{% block content %}
-    <div class="flex flex-row">
-        <form class="max-w-full my-28 bg-white shadow-md rounded-md p-8 mx-auto" method="POST">
-            <div class="text-2xl text-center font-bold mb-6 drop-shadow-md">Predict Salary Rate</div>
-            <div class="mb-5">
-                <label for="major" class="block mb-2 text-sm font-medium text-gray-900 dark:text-white">
-                    Select a Major
-                </label>
-                <select id="major" name="Major" class="bg-gray-50 border border-gray-300 text-gray-900 text-sm rounded-lg focus:ring-blue-500 focus:border-blue-500 block w-full p-2.5 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-blue-500 dark:focus:border-blue-500">
-                    {% for major in majors %}
-                        <option value="{{ major }}">{{ major }}</option>
-                    {% endfor %}
-                </select>
-            </div>
-            <div class="mb-5">
-                <label for="P25th" class="block mb-2 text-sm font-medium text-gray-900 dark:text-white">
-                    Average salary per year below average
-                </label>
-                <input type="number" id="P25th" name="P2th" class="bg-gray-50 border border-gray-300 text-gray-900 text-sm rounded-lg focus:ring-blue-500 focus:border-blue-500 block w-full p-2.5 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-blue-500 dark:focus:border-blue-500" required />
-            </div>
-            <div class="mb-5">
-                <label for="P75th" class="block mb-2 text-sm font-medium text-gray-900 dark:text-white">
-                    Average salary per year above average
-                </label>
-                <input type="number" id="P75th" name="P75th" class="bg-gray-50 border border-gray-300 text-gray-900 text-sm rounded-lg focus:ring-blue-500 focus:border-blue-500 block w-full p-2.5 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-blue-500 dark:focus:border-blue-500" required />
-            </div>
-            <button type="submit" class="text-white bg-blue-700 hover:bg-blue-800 focus:ring-4 focus:outline-none focus:ring-blue-300 font-medium rounded-lg text-sm w-full sm:w-auto px-5 py-2.5 text-center dark:bg-blue-600 dark:hover:bg-blue-700 dark:focus:ring-blue-800 transition-all">
-                Process
-            </button>
-        </form>
-        
-        {% if prediction_text %}
-            <div class="max-w-prose p-6 mx-auto h-fit mt-28 bg-white border border-gray-200 rounded-lg shadow dark:bg-gray-800 dark:border-gray-700">
-                <a href="#">
-                    <h5 class="mb-2 text-2xl font-bold tracking-tight text-gray-900 dark:text-white">Hasil Prediksi Salary Rate</h5>
-                </a>
-                <p class="mb-3 font-normal text-gray-700 dark:text-gray-400">{{ prediction_text }}</p>
-                <p class="mb-3 font-normal text-gray-700 dark:text-gray-400">
-                    You may be wondering how we can determine whether the major you enter and process is advised or not. In order to decide whether or not the major is advised, we compared the processed data with the average pay rate per year, which is $59,384 in USD.
-                </p>
-                <a href="/chart" class="inline-flex items-center px-3 py-2 text-sm font-medium text-center text-white bg-blue-700 rounded-lg hover:bg-blue-800 focus:ring-4 focus:outline-none focus:ring-blue-300 dark:bg-blue-600 dark:hover:bg-blue-700 dark:focus:ring-blue-800">
-                    Read more
-                    <svg class="rtl:rotate-180 w-3.5 h-3.5 ms-2" aria-hidden="true" xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 14 10">
-                        <path stroke="currentColor" stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M1 5h12m0 0L9 1m4 4L9 9"/>
-                    </svg>
-                </a>
-            </div>
-        {% endif %}
-    </div>
-{% endblock %}
